Added Exception Handling Shifted From Hibernate Daos to JpaRepository Made the application more interactive Added Beautification

+++ project/src/main/java/com/springboot/project/controller/AdminController.java
@@ -3,8 +3,9 @@ package com.springboot.project.controller;
 
 import com.springboot.project.DTO.BookDTO;
 import com.springboot.project.entity.Book;
+import com.springboot.project.entity.BookStatus;
+
 import com.springboot.project.service.BookService;
-import com.springboot.project.service.MemberService;
 import com.springboot.project.service.TransactionService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,22 +18,21 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 public class AdminController {
 
-    MemberService memberService;
-    BookService bookService;
+    private final BookService bookService;
 
-    TransactionService transactionService;
+    private final TransactionService transactionService;
     @Autowired
-    public AdminController(MemberService memberService, BookService bookService, TransactionService transactionService) {
-        this.memberService = memberService;
+    public AdminController(BookService bookService, TransactionService transactionService) {
         this.bookService = bookService;
         this.transactionService = transactionService;
     }
 
 
     @GetMapping("/home")
-    public String Home(Model model){
-        model.addAttribute("Books", bookService.findAll());
-        model.addAttribute("Transactions", transactionService.findAllWithFines(transactionService.findAll()));
+    public String home(@RequestParam(value = "errorMessage", required = false) String error, Model model){
+        model.addAttribute("ErrorMessage", error);
+        model.addAttribute("Books", bookService.findAllBooks());
+        model.addAttribute("Transactions", transactionService.findFinesForGivenTransactions(transactionService.findAll()));
         return "admin-home";
     }
 
@@ -42,40 +42,54 @@ public class AdminController {
         return "book-form";
     }
     @PostMapping("/book")
-    public String saveBook(@Valid @ModelAttribute("Book") BookDTO bookDTO, BindingResult result, Model model){
+    public String saveAndUpdateBook(@Valid @ModelAttribute("Book") BookDTO bookDTO, BindingResult result, Model model){
 
         if(result.hasErrors()){
             model.addAttribute("Book", bookDTO);
             return "book-form";
         }
         Book book;
-        System.out.println(bookDTO.getIsbn());
+
         if(bookDTO.getIsbn() > -1){
-            book = bookService.Search(bookDTO.getIsbn());
+            book = bookService.SearchByUserQuery(bookDTO.getIsbn());
+            if(( book == null) || book.getStatus()== BookStatus.REMOVED) return "redirect:/admin/home?errorMessage=Error: Invalid ISBN. ";
             book.setAuthor(bookDTO.getAuthor());
             book.setEdition(bookDTO.getEdition());
             book.setTitle(bookDTO.getTitle());
             book.setPublisher(bookDTO.getPublisher());
         } else {
             book = new Book(bookDTO.getTitle(), bookDTO.getAuthor(), bookDTO.getPublisher(), bookDTO.getEdition());
-            System.out.println(book.getIsbn());
         }
-        bookService.save(book);
-        return "redirect:/admin/home";
+        try {
+            bookService.save(book);
+        } catch (Exception e){
+            return "redirect:/admin/home?errorMessage=Database Error: Book can't be saved at this time";
+        }
+        return "redirect:/admin/home?success=true";
     }
 
     @GetMapping("/show-update-book-form")
     public String updateBookPage(@RequestParam("bookISBN") String isbn, Model model){
-        var book = bookService.Search(Integer.parseInt(isbn));
-        var bookDTO = new BookDTO(book.getIsbn() ,book.getTitle(), book.getAuthor(), book.getPublisher(), book.getEdition());
-        model.addAttribute("Book", bookDTO);
+        try{
+            var book = bookService.SearchByUserQuery(Integer.parseInt(isbn));
+            var bookDTO = new BookDTO(book.getIsbn(), book.getTitle(), book.getAuthor(), book.getPublisher(), book.getEdition());
+            model.addAttribute("Book", bookDTO);
+        }catch (NullPointerException e){
+            return "redirect:/admin/home?errorMessage=Error: Invalid ISBN. ";
+        }
         return "book-form";
     }
 
     @GetMapping("/delete")
     public String deleteBook(@RequestParam("bookISBN") int isbn){
-        bookService.delete(isbn);
-        return "redirect:/admin/home";
+        var book = bookService.SearchByUserQuery(isbn);
+        if(( book == null) || book.getStatus()== BookStatus.REMOVED) return "redirect:/admin/home?errorMessage=Error: Invalid ISBN. ";
+        StringBuilder errorMessages = new StringBuilder();
+        try{
+            bookService.delete(isbn);
+            return "redirect:/admin/home?success=true";
+        } catch(Exception e){
+            return "redirect:/admin/home?errorMessage=DatabaseError: The book couldn't be deleted. ";
+        }
     }
-
 }
